refactor: extract pipeline/hash.rs
Move SSOT content-hashing helpers (hash_item_dir, collect_files,
hash_source_items) out of pipeline/mod.rs into a focused submodule.

hash_item_dir keeps its pub(crate) visibility (re-exported by mod.rs)
since lockfile and other sibling modules call it. hash_source_items
stays pub(super) — used only by update logic still in mod.rs.

Refs #180
